fix: cross-project listener management to avoid bugs
