+++ pluginVersion.txt
@@ -1,1 +1,1 @@
-1.9.9
+1.9.9b
